+++ client/app/Domain/Entity/Encoder.go
@@ -1,10 +1,8 @@
 package Entity
 
-
 type Encoder interface {
 	Watchdog()
-	ResetPosition() 
-	GetPosition() int16
+	ResetPosition()
+	GetPosition() int64
 	TearDown()
 }
-
